@@ -1,8 +1,10 @@
+#!/usr/bin/env python
+"""pyiCloud setup."""
+
 from setuptools import setup, find_packages
-from codecs import open
 
 REPO_URL = "https://github.com/picklepete/pyicloud"
-VERSION = "0.10.2"
+VERSION = "1.0.0"
 
 with open("requirements.txt") as f:
     required = f.read().splitlines()
@@ -18,24 +20,23 @@ setup(
     description="PyiCloud is a module which allows pythonistas to interact with iCloud webservices.",
     long_description=long_description,
     maintainer="The PyiCloud Authors",
-    maintainer_email=" ",
     packages=find_packages(include=["pyicloud*"]),
     install_requires=required,
-    python_requires=">=2.7, !=3.0.*, !=3.1.*, !=3.2.*",
+    python_requires=">=3.7",
     license="MIT",
     classifiers=[
+        "Development Status :: 5 - Production/Stable",
         "Intended Audience :: Developers",
         "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
         "Programming Language :: Python",
-        "Programming Language :: Python :: 2.7",
+        "Programming Language :: Python :: 3 :: Only",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.3",
-        "Programming Language :: Python :: 3.4",
-        "Programming Language :: Python :: 3.5",
-        "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Topic :: Software Development :: Libraries",
     ],
     entry_points={"console_scripts": ["icloud = pyicloud.cmdline:main"]},
     keywords=["icloud", "find-my-iphone"],
