@@ -1,12 +1,10 @@
 """Account service."""
-from __future__ import division
-from six import PY2, python_2_unicode_compatible
 from collections import OrderedDict
 
 from pyicloud.utils import underscore_to_camelcase
 
 
-class AccountService(object):
+class AccountService:
     """The 'Account' iCloud service."""
 
     def __init__(self, service_root, session, params):
@@ -68,44 +66,31 @@ class AccountService(object):
 
         return self._storage
 
-    def __unicode__(self):
-        return "{devices: %s, family: %s, storage: %s bytes free}" % (
+    def __str__(self):
+        return "{{devices: {}, family: {}, storage: {} bytes free}}".format(
             len(self.devices),
             len(self.family),
             self.storage.usage.available_storage_in_bytes,
         )
 
-    def __str__(self):
-        as_unicode = self.__unicode__()
-        if PY2:
-            return as_unicode.encode("utf-8", "ignore")
-        return as_unicode
-
     def __repr__(self):
-        return "<%s: %s>" % (type(self).__name__, str(self))
+        return f"<{type(self).__name__}: {self}>"
 
 
-@python_2_unicode_compatible
 class AccountDevice(dict):
     """Account device."""
 
     def __getattr__(self, key):
         return self[underscore_to_camelcase(key)]
 
-    def __unicode__(self):
-        return "{model: %s, name: %s}" % (self.model_display_name, self.name)
-
     def __str__(self):
-        as_unicode = self.__unicode__()
-        if PY2:
-            return as_unicode.encode("utf-8", "ignore")
-        return as_unicode
+        return f"{{model: {self.model_display_name}, name: {self.name}}}"
 
     def __repr__(self):
-        return "<%s: %s>" % (type(self).__name__, str(self))
+        return f"<{type(self).__name__}: {self}>"
 
 
-class FamilyMember(object):
+class FamilyMember:
     """A family member."""
 
     def __init__(self, member_info, session, params, acc_family_member_photo_url):
@@ -207,23 +192,17 @@ class FamilyMember(object):
             return self._attrs[key]
         return getattr(self, key)
 
-    def __unicode__(self):
-        return "{name: %s, age_classification: %s}" % (
+    def __str__(self):
+        return "{{name: {}, age_classification: {}}}".format(
             self.full_name,
             self.age_classification,
         )
 
-    def __str__(self):
-        as_unicode = self.__unicode__()
-        if PY2:
-            return as_unicode.encode("utf-8", "ignore")
-        return as_unicode
-
     def __repr__(self):
-        return "<%s: %s>" % (type(self).__name__, str(self))
+        return f"<{type(self).__name__}: {self}>"
 
 
-class AccountStorageUsageForMedia(object):
+class AccountStorageUsageForMedia:
     """Storage used for a specific media type into the account."""
 
     def __init__(self, usage_data):
@@ -249,20 +228,14 @@ class AccountStorageUsageForMedia(object):
         """Gets the usage in bytes."""
         return self.usage_data["usageInBytes"]
 
-    def __unicode__(self):
-        return "{key: %s, usage: %s bytes}" % (self.key, self.usage_in_bytes)
-
     def __str__(self):
-        as_unicode = self.__unicode__()
-        if PY2:
-            return as_unicode.encode("utf-8", "ignore")
-        return as_unicode
+        return f"{{key: {self.key}, usage: {self.usage_in_bytes} bytes}}"
 
     def __repr__(self):
-        return "<%s: %s>" % (type(self).__name__, str(self))
+        return f"<{type(self).__name__}: {self}>"
 
 
-class AccountStorageUsage(object):
+class AccountStorageUsage:
     """Storage used for a specific media type into the account."""
 
     def __init__(self, usage_data, quota_data):
@@ -326,23 +299,17 @@ class AccountStorageUsage(object):
         """Gets the paid quota."""
         return self.quota_data["paidQuota"]
 
-    def __unicode__(self):
-        return "%s%% used of %s bytes" % (
+    def __str__(self):
+        return "{}% used of {} bytes".format(
             self.used_storage_in_percent,
             self.total_storage_in_bytes,
         )
 
-    def __str__(self):
-        as_unicode = self.__unicode__()
-        if PY2:
-            return as_unicode.encode("utf-8", "ignore")
-        return as_unicode
-
     def __repr__(self):
-        return "<%s: %s>" % (type(self).__name__, str(self))
+        return f"<{type(self).__name__}: {self}>"
 
 
-class AccountStorage(object):
+class AccountStorage:
     """Storage of the account."""
 
     def __init__(self, storage_data):
@@ -356,14 +323,8 @@ class AccountStorage(object):
                 usage_media
             )
 
-    def __unicode__(self):
-        return "{usage: %s, usages_by_media: %s}" % (self.usage, self.usages_by_media)
-
     def __str__(self):
-        as_unicode = self.__unicode__()
-        if PY2:
-            return as_unicode.encode("utf-8", "ignore")
-        return as_unicode
+        return f"{{usage: {self.usage}, usages_by_media: {self.usages_by_media}}}"
 
     def __repr__(self):
-        return "<%s: %s>" % (type(self).__name__, str(self))
+        return f"<{type(self).__name__}: {self}>"
