@@ -1,16 +1,21 @@
 """Drive service."""
 from datetime import datetime, timedelta
 import json
+import logging
 import io
 import mimetypes
 import os
 import time
 from re import search
 from requests import Response
-from six import PY2
+
+from pyicloud.exceptions import PyiCloudAPIResponseException
 
 
-class DriveService(object):
+LOGGER = logging.getLogger(__name__)
+
+
+class DriveService:
     """The 'Drive' iCloud service."""
 
     def __init__(self, service_root, document_root, session, params):
@@ -43,6 +48,7 @@ class DriveService(object):
                 ]
             ),
         )
+        self._raise_if_error(request)
         return request.json()[0]
 
     def get_file(self, file_id, **kwargs):
@@ -53,16 +59,22 @@ class DriveService(object):
             self._document_root + "/ws/com.apple.CloudDocs/download/by_id",
             params=file_params,
         )
-        if not response.ok:
-            return None
-        url = response.json()["data_token"]["url"]
-        return self.session.get(url, params=self.params, **kwargs)
+        self._raise_if_error(response)
+        response_json = response.json()
+        package_token = response_json.get("package_token")
+        data_token = response_json.get("data_token")
+        if data_token and data_token.get("url"):
+            return self.session.get(data_token["url"], params=self.params, **kwargs)
+        if package_token and package_token.get("url"):
+            return self.session.get(package_token["url"], params=self.params, **kwargs)
+        raise KeyError("'data_token' nor 'package_token'")
 
     def get_app_data(self):
         """Returns the app library (previously ubiquity)."""
         request = self.session.get(
             self._service_root + "/retrieveAppLibraries", params=self.params
         )
+        self._raise_if_error(request)
         return request.json()["items"]
 
     def _get_upload_contentws_url(self, file_object):
@@ -93,8 +105,7 @@ class DriveService(object):
                 }
             ),
         )
-        if not request.ok:
-            return None
+        self._raise_if_error(request)
         return (request.json()[0]["document_id"], request.json()[0]["url"])
 
     def _update_contentws(self, folder_id, sf_info, document_id, file_object):
@@ -108,7 +119,10 @@ class DriveService(object):
             "command": "add_file",
             "create_short_guid": True,
             "document_id": document_id,
-            "path": {"starting_document_id": folder_id, "path": file_object.name,},
+            "path": {
+                "starting_document_id": folder_id,
+                "path": file_object.name,
+            },
             "allow_conflict": True,
             "file_flags": {
                 "is_writable": True,
@@ -129,8 +143,7 @@ class DriveService(object):
             headers={"Content-Type": "text/plain"},
             data=json.dumps(data),
         )
-        if not request.ok:
-            return None
+        self._raise_if_error(request)
         return request.json()
 
     def send_file(self, folder_id, file_object):
@@ -138,10 +151,8 @@ class DriveService(object):
         document_id, content_url = self._get_upload_contentws_url(file_object)
 
         request = self.session.post(content_url, files={file_object.name: file_object})
-        if not request.ok:
-            return None
+        self._raise_if_error(request)
         content_response = request.json()["singleFile"]
-
         self._update_contentws(folder_id, content_response, document_id, file_object)
 
     def create_folders(self, parent, name):
@@ -153,10 +164,16 @@ class DriveService(object):
             data=json.dumps(
                 {
                     "destinationDrivewsId": parent,
-                    "folders": [{"clientId": self.params["clientId"], "name": name,}],
+                    "folders": [
+                        {
+                            "clientId": self.params["clientId"],
+                            "name": name,
+                        }
+                    ],
                 }
             ),
         )
+        self._raise_if_error(request)
         return request.json()
 
     def rename_items(self, node_id, etag, name):
@@ -165,9 +182,18 @@ class DriveService(object):
             self._service_root + "/renameItems",
             params=self.params,
             data=json.dumps(
-                {"items": [{"drivewsid": node_id, "etag": etag, "name": name,}],}
+                {
+                    "items": [
+                        {
+                            "drivewsid": node_id,
+                            "etag": etag,
+                            "name": name,
+                        }
+                    ],
+                }
             ),
         )
+        self._raise_if_error(request)
         return request.json()
 
     def move_items_to_trash(self, node_id, etag):
@@ -187,6 +213,7 @@ class DriveService(object):
                 }
             ),
         )
+        self._raise_if_error(request)
         return request.json()
 
     @property
@@ -202,8 +229,16 @@ class DriveService(object):
     def __getitem__(self, key):
         return self.root[key]
 
+    def _raise_if_error(self, response):  # pylint: disable=no-self-use
+        if not response.ok:
+            api_error = PyiCloudAPIResponseException(
+                response.reason, response.status_code
+            )
+            LOGGER.error(api_error)
+            raise api_error
 
-class DriveNode(object):
+
+class DriveNode:
     """Drive node."""
 
     def __init__(self, conn, data):
@@ -215,7 +250,7 @@ class DriveNode(object):
     def name(self):
         """Gets the node name."""
         if "extension" in self.data:
-            return "%s.%s" % (self.data["name"], self.data["extension"])
+            return "{}.{}".format(self.data["name"], self.data["extension"])
         return self.data["name"]
 
     @property
@@ -270,7 +305,7 @@ class DriveNode(object):
         return self.connection.get_file(self.data["docwsid"], **kwargs)
 
     def upload(self, file_object, **kwargs):
-        """"Upload a new file."""
+        """Upload a new file."""
         return self.connection.send_file(self.data["docwsid"], file_object, **kwargs)
 
     def dir(self):
@@ -304,20 +339,14 @@ class DriveNode(object):
     def __getitem__(self, key):
         try:
             return self.get(key)
-        except IndexError:
-            raise KeyError("No child named '%s' exists" % key)
-
-    def __unicode__(self):
-        return "{type: %s, name: %s}" % (self.type, self.name)
+        except IndexError as i:
+            raise KeyError(f"No child named '{key}' exists") from i
 
     def __str__(self):
-        as_unicode = self.__unicode__()
-        if PY2:
-            return as_unicode.encode("utf-8", "ignore")
-        return as_unicode
+        return rf"\{type: {self.type}, name: {self.name}\}"
 
     def __repr__(self):
-        return "<%s: %s>" % (type(self).__name__, str(self))
+        return f"<{type(self).__name__}: {str(self)}>"
 
 
 def _date_to_utc(date):
