@@ -1,12 +1,10 @@
 """Find my iPhone service."""
 import json
 
-from six import PY2, text_type
-
 from pyicloud.exceptions import PyiCloudNoDevicesException
 
 
-class FindMyiPhoneServiceManager(object):
+class FindMyiPhoneServiceManager:
     """The 'Find my iPhone' iCloud service
 
     This connects to iCloud and return phone data including the near-realtime
@@ -69,29 +67,20 @@ class FindMyiPhoneServiceManager(object):
 
     def __getitem__(self, key):
         if isinstance(key, int):
-            if PY2:
-                key = self.keys()[key]
-            else:
-                key = list(self.keys())[key]
+            key = list(self.keys())[key]
         return self._devices[key]
 
     def __getattr__(self, attr):
         return getattr(self._devices, attr)
 
-    def __unicode__(self):
-        return text_type(self._devices)
-
     def __str__(self):
-        as_unicode = self.__unicode__()
-        if PY2:
-            return as_unicode.encode("utf-8", "ignore")
-        return as_unicode
+        return f"{self._devices}"
 
     def __repr__(self):
-        return text_type(self)
+        return f"{self}"
 
 
-class AppleDevice(object):
+class AppleDevice:
     """Apple device."""
 
     def __init__(
@@ -200,16 +189,8 @@ class AppleDevice(object):
     def __getattr__(self, attr):
         return getattr(self.content, attr)
 
-    def __unicode__(self):
-        display_name = self["deviceDisplayName"]
-        name = self["name"]
-        return "%s: %s" % (display_name, name)
-
     def __str__(self):
-        as_unicode = self.__unicode__()
-        if PY2:
-            return as_unicode.encode("utf-8", "ignore")
-        return as_unicode
+        return f"{self['deviceDisplayName']}: {self['name']}"
 
     def __repr__(self):
-        return "<AppleDevice(%s)>" % str(self)
+        return f"<AppleDevice({self})>"
