@@ -1,18 +1,17 @@
-# -*- coding: utf-8 -*-
 """Account service tests."""
 from unittest import TestCase
-from six import PY3
 
 from . import PyiCloudServiceMock
 from .const import AUTHENTICATED_USER, VALID_PASSWORD
 
 
 class AccountServiceTest(TestCase):
-    """"Account service tests"""
+    """Account service tests."""
 
     service = None
 
     def setUp(self):
+        """Set up tests."""
         self.service = PyiCloudServiceMock(AUTHENTICATED_USER, VALID_PASSWORD).account
 
     def test_repr(self):
@@ -42,8 +41,7 @@ class AccountServiceTest(TestCase):
             assert device["modelSmallPhotoURL1x"]
             assert device["modelDisplayName"]
             # fmt: off
-            if PY3:
-                assert repr(device) == "<AccountDevice: {model: "+device.model_display_name+", name: "+device.name+"}>"
+            assert repr(device) == "<AccountDevice: {model: "+device.model_display_name+", name: "+device.name+"}>"
             # fmt: on
 
     def test_family(self):
@@ -72,8 +70,7 @@ class AccountServiceTest(TestCase):
         """Tests storage."""
         assert self.service.storage
         # fmt: off
-        if PY3:
-            assert repr(self.service.storage) == "<AccountStorage: {usage: 43.75% used of 5368709120 bytes, usages_by_media: OrderedDict([('photos', <AccountStorageUsageForMedia: {key: photos, usage: 0 bytes}>), ('backup', <AccountStorageUsageForMedia: {key: backup, usage: 799008186 bytes}>), ('docs', <AccountStorageUsageForMedia: {key: docs, usage: 449092146 bytes}>), ('mail', <AccountStorageUsageForMedia: {key: mail, usage: 1101522944 bytes}>)])}>"
+        assert repr(self.service.storage) == "<AccountStorage: {usage: 43.75% used of 5368709120 bytes, usages_by_media: OrderedDict([('photos', <AccountStorageUsageForMedia: {key: photos, usage: 0 bytes}>), ('backup', <AccountStorageUsageForMedia: {key: backup, usage: 799008186 bytes}>), ('docs', <AccountStorageUsageForMedia: {key: docs, usage: 449092146 bytes}>), ('mail', <AccountStorageUsageForMedia: {key: mail, usage: 1101522944 bytes}>)])}>"
         # fmt: on
 
     def test_storage_usage(self):
