@@ -1,20 +1,15 @@
-# -*- coding: utf-8 -*-
 """Cmdline tests."""
-from pyicloud import cmdline
-from . import PyiCloudServiceMock
-from .const import AUTHENTICATED_USER, REQUIRES_2FA_USER, VALID_PASSWORD, VALID_2FA_CODE
-from .const_findmyiphone import FMI_FAMILY_WORKING
-
 import os
-from six import PY2
 import pickle
-import pytest
 from unittest import TestCase
+from unittest.mock import patch
 
-if PY2:
-    from mock import patch
-else:
-    from unittest.mock import patch  # pylint: disable=no-name-in-module,import-error
+import pytest
+from pyicloud import cmdline
+
+from . import PyiCloudServiceMock
+from .const import AUTHENTICATED_USER, REQUIRES_2FA_USER, VALID_2FA_CODE, VALID_PASSWORD
+from .const_findmyiphone import FMI_FAMILY_WORKING
 
 
 class TestCmdline(TestCase):
@@ -23,6 +18,7 @@ class TestCmdline(TestCase):
     main = None
 
     def setUp(self):
+        """Set up tests."""
         cmdline.PyiCloudService = PyiCloudServiceMock
         self.main = cmdline.main
 
