@@ -1,16 +1,17 @@
-# -*- coding: utf-8 -*-
 """Find My iPhone service tests."""
 from unittest import TestCase
+
 from . import PyiCloudServiceMock
 from .const import AUTHENTICATED_USER, VALID_PASSWORD
 
 
 class FindMyiPhoneServiceTest(TestCase):
-    """"Find My iPhone service tests"""
+    """Find My iPhone service tests."""
 
     service = None
 
     def setUp(self):
+        """Set up tests."""
         self.service = PyiCloudServiceMock(AUTHENTICATED_USER, VALID_PASSWORD)
 
     def test_devices(self):
